requirements.txt
redis>=8.0
redisvl>=0.0.3

psycopg[binary]>=3.0
psycopg_pool>=3.1.1

numpy